% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{.cleanCheckpoints}
\alias{.cleanCheckpoints}
\title{Remove checkpoint files and directory}
\usage{
.cleanCheckpoints(checkpoint_dir)
}
\description{
Remove checkpoint files and directory
}
\keyword{internal}
