% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{.collapseSegments}
\alias{.collapseSegments}
\title{Collapse adjacent CDS segments back to original exon boundaries}
\usage{
.collapseSegments(track_data)
}
\description{
Collapse adjacent CDS segments back to original exon boundaries
}
\keyword{internal}
