% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uorf.R
\name{.compute5UtrLength}
\alias{.compute5UtrLength}
\title{Compute 5'UTR Length in Transcript Coordinates}
\usage{
.compute5UtrLength(strand, exon_starts, exon_ends, cds_start, cds_stop)
}
\arguments{
\item{strand}{Character; "+" or "-".}

\item{exon_starts}{Integer vector of exon start coordinates.}

\item{exon_ends}{Integer vector of exon end coordinates.}

\item{cds_start}{Integer; min genomic CDS coordinate.}

\item{cds_stop}{Integer; max genomic CDS coordinate.}
}
\value{
Integer 5'UTR length in nucleotides.
}
\description{
Compute 5'UTR Length in Transcript Coordinates
}
\keyword{internal}
