% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-detection.R
\name{.computeBoundaryCoords}
\alias{.computeBoundaryCoords}
\title{Compute five_prime/three_prime coordinates for boundary events}
\usage{
.computeBoundaryCoords(event_type, direction, strand, ref_exon, comp_exon)
}
\description{
Compute five_prime/three_prime coordinates for boundary events
}
\keyword{internal}
