% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein-mapping.R
\name{.computeCdsLength}
\alias{.computeCdsLength}
\title{Compute Total CDS Length in Nucleotides}
\usage{
.computeCdsLength(exon_starts, exon_ends, cds_start, cds_stop)
}
\arguments{
\item{exon_starts}{Integer vector of exon starts.}

\item{exon_ends}{Integer vector of exon ends.}

\item{cds_start}{Integer; min genomic CDS coordinate.}

\item{cds_stop}{Integer; max genomic CDS coordinate.}
}
\value{
Integer total CDS length in nucleotides.
}
\description{
Sums up exonic bases that fall within the CDS boundaries.
}
\keyword{internal}
