% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{.computeSharedJunctions}
\alias{.computeSharedJunctions}
\title{Compute shared junctions between two isoforms (Jaccard)}
\usage{
.computeSharedJunctions(ref_junctions, comp_junctions)
}
\description{
Compute shared junctions between two isoforms (Jaccard)
}
\keyword{internal}
