% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uorf.R
\name{.computeTranscriptLength}
\alias{.computeTranscriptLength}
\title{Compute Total Transcript Length}
\usage{
.computeTranscriptLength(exon_starts, exon_ends)
}
\arguments{
\item{exon_starts}{Integer vector of exon start coordinates.}

\item{exon_ends}{Integer vector of exon end coordinates.}
}
\value{
Integer total transcript length in nucleotides.
}
\description{
Compute Total Transcript Length
}
\keyword{internal}
