% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{.computeUeComparison}
\alias{.computeUeComparison}
\title{Compare union exon coverage between two isoforms}
\usage{
.computeUeComparison(ref_id, comp_id, isoform_union_mapping)
}
\description{
Compare union exon coverage between two isoforms
}
\keyword{internal}
