% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-detection.R
\name{.detectSharedBoundaryEvent}
\alias{.detectSharedBoundaryEvent}
\title{Detect shared-boundary and dual-boundary splicing events}
\usage{
.detectSharedBoundaryEvent(
  exon_ref,
  exon_comp,
  strand,
  is_first_exon = FALSE,
  is_last_exon = FALSE,
  is_first_exon_ref = FALSE,
  is_last_exon_ref = FALSE,
  is_first_exon_comp = FALSE,
  is_last_exon_comp = FALSE,
  splice_site_threshold = 100L
)
}
\description{
Detect shared-boundary and dual-boundary splicing events
}
\keyword{internal}
