% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uorf.R
\name{.detectUorfsOrfik}
\alias{.detectUorfsOrfik}
\title{ORFik-based uORF Detection (Internal)}
\usage{
.detectUorfsOrfik(
  structures,
  cds_metadata,
  sequences,
  min_length_codons,
  kozak,
  verbose
)
}
\description{
ORFik-based uORF Detection (Internal)
}
\keyword{internal}
