% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein-utils.R
\name{.digestTrypsin}
\alias{.digestTrypsin}
\title{Trypsin In Silico Digestion}
\usage{
.digestTrypsin(seq)
}
\arguments{
\item{seq}{Character; single amino acid sequence.}
}
\value{
Character vector of peptide fragments.
}
\description{
Cleaves after K or R, except when followed by P.
}
\keyword{internal}
