% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{.expandStructures}
\alias{.expandStructures}
\title{Expand nested structures to per-exon rows}
\usage{
.expandStructures(structures)
}
\description{
Expand nested structures to per-exon rows
}
\keyword{internal}
