% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptc-attribution.R
\name{.findFirstStop}
\alias{.findFirstStop}
\title{Find first in-frame stop codon in a sequence}
\usage{
.findFirstStop(seq_str, atg_pos, stop_codons)
}
\description{
Find first in-frame stop codon in a sequence
}
\keyword{internal}
