% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptc-attribution.R
\name{.getJunctionPositions}
\alias{.getJunctionPositions}
\title{Get junction positions in transcript space}
\usage{
.getJunctionPositions(exon_starts, exon_ends, strand)
}
\description{
Get junction positions in transcript space
}
\keyword{internal}
