% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptc.R
\name{.getStartCodonPos}
\alias{.getStartCodonPos}
\title{Get Start Codon Position}
\usage{
.getStartCodonPos(exons, strand)
}
\arguments{
\item{exons}{Data frame with cds_exon_start, cds_exon_end for one isoform.}

\item{strand}{"+" or "-".}
}
\value{
Integer genomic position.
}
\description{
Returns the genomic position of the 5'-most CDS nucleotide for an isoform.
}
\keyword{internal}
