% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptc.R
\name{.intersectCdsIntervals}
\alias{.intersectCdsIntervals}
\title{Intersect Two Sets of Genomic Intervals}
\usage{
.intersectCdsIntervals(a, b)
}
\arguments{
\item{a}{Data frame with columns \code{start} and \code{end} (integer).}

\item{b}{Data frame with columns \code{start} and \code{end} (integer).}
}
\value{
Data frame with columns \code{start} and \code{end} for the
intersection regions.
}
\description{
Computes the genomic intersection of two sets of genomic intervals.
Both inputs are assumed to be non-overlapping within each set.
}
\keyword{internal}
