% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptc-attribution.R
\name{.isCodonExonic}
\alias{.isCodonExonic}
\title{Check if a 3-nucleotide codon is exonic}
\usage{
.isCodonExonic(pos, exon_starts, exon_ends, strand)
}
\description{
Check if a 3-nucleotide codon is exonic
}
\keyword{internal}
