% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{.makeIntrons}
\alias{.makeIntrons}
\title{Build intron connector segments for all tracks}
\usage{
.makeIntrons(all_exons)
}
\description{
Build intron connector segments for all tracks
}
\keyword{internal}
