% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruction.R
\name{.modifyExonBoundary}
\alias{.modifyExonBoundary}
\title{Modify exon boundary for splice site events}
\usage{
.modifyExonBoundary(exons, event)
}
\description{
Modify exon boundary for splice site events
}
\keyword{internal}
