% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruction.R
\name{.modifyTerminalExon}
\alias{.modifyTerminalExon}
\title{Modify terminal exon for Alt_TSS/Alt_TES events}
\usage{
.modifyTerminalExon(exons, event)
}
\description{
Modify terminal exon for Alt_TSS/Alt_TES events
}
\keyword{internal}
