% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrence.R
\name{.runCooccurrenceTests}
\alias{.runCooccurrenceTests}
\title{Run pairwise co-occurrence Fisher's tests}
\usage{
.runCooccurrenceTests(data, pairs, event_flags, event_names)
}
\description{
Run pairwise co-occurrence Fisher's tests
}
\keyword{internal}
