% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein-utils.R
\name{.writeSequencesToFasta}
\alias{.writeSequencesToFasta}
\title{Write Protein Sequences to a Temporary FASTA File}
\usage{
.writeSequencesToFasta(sequences, temp_dir)
}
\arguments{
\item{sequences}{Named character vector of AA sequences.}

\item{temp_dir}{Directory for the temporary file.}
}
\value{
Path to the written FASTA file.
}
\description{
Write Protein Sequences to a Temporary FASTA File
}
\keyword{internal}
