% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{mapEventsToRegions}
\alias{mapEventsToRegions}
\title{Map Events to Genomic Regions}
\usage{
mapEventsToRegions(profiles, annotated_ue, cds_metadata)
}
\arguments{
\item{profiles}{A tibble from \code{\link{buildProfiles}()} with
\code{detailed_events}.}

\item{annotated_ue}{A tibble from \code{\link{annotateRegionTypes}()}
with columns: \code{isoform_id}, \code{union_exon_id},
\code{isoform_exon_start}, \code{isoform_exon_end},
\code{region_type}.}

\item{cds_metadata}{A tibble from \code{\link{extractCdsAnnotations}()}
with \code{isoform_id}, \code{cds_start}, \code{cds_stop}.}
}
\value{
A tibble with columns: \code{event_row_id}, \code{gene_id},
\code{reference_isoform_id}, \code{comparator_isoform_id},
\code{event_type}, \code{direction}, \code{genomic_start},
\code{genomic_end}, \code{region_type}.
}
\description{
Maps each splicing event to overlapping genomic regions (5'UTR, CDS,
3'UTR, etc.) using annotated union exon mappings. Events can map to
multiple region types when they span region boundaries.
}
\examples{
gtf_path <- system.file("extdata", "example_small.gtf", package = "Isopair")
structures <- parseIsoformStructures(gtf_path, verbose = FALSE)
ue <- buildUnionExons(structures, verbose = FALSE)
cds <- extractCdsAnnotations(gtf_path, structures$isoform_id,
  verbose = FALSE)
annotated <- annotateRegionTypes(ue$isoform_union_mapping, cds)
data(example_profiles)
er <- mapEventsToRegions(example_profiles, annotated, cds)
head(er)
}
