% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{summarizeOrfImpact}
\alias{summarizeOrfImpact}
\title{Summarize ORF Impact of Terminal Events}
\usage{
summarizeOrfImpact(event_regions, cds_metadata)
}
\arguments{
\item{event_regions}{A tibble from \code{\link{mapEventsToRegions}()}.}

\item{cds_metadata}{A tibble from \code{\link{extractCdsAnnotations}()}.}
}
\value{
A tibble with columns: \code{event_type}, \code{n_events},
\code{n_affecting_orf_start}, \code{pct_affecting_orf_start},
\code{n_affecting_orf_stop}, \code{pct_affecting_orf_stop}.
}
\description{
For Alt_TSS and Alt_TES events, determines what fraction overlap the ORF
start or stop codon in the reference isoform.
}
\examples{
gtf_path <- system.file("extdata", "example_small.gtf", package = "Isopair")
structures <- parseIsoformStructures(gtf_path, verbose = FALSE)
ue <- buildUnionExons(structures, verbose = FALSE)
cds <- extractCdsAnnotations(gtf_path, structures$isoform_id,
  verbose = FALSE)
annotated <- annotateRegionTypes(ue$isoform_union_mapping, cds)
data(example_profiles)
er <- mapEventsToRegions(example_profiles, annotated, cds)
orf_impact <- summarizeOrfImpact(er, cds)
orf_impact
}
