% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{summarizeProfiles}
\alias{summarizeProfiles}
\title{Summarize Profiles}
\usage{
summarizeProfiles(profiles)
}
\arguments{
\item{profiles}{A tibble from \code{\link[=buildProfiles]{buildProfiles()}}.}
}
\value{
A named list of summary statistics.
}
\description{
Compute summary statistics across a set of splicing choice profiles.
}
\examples{
data(example_profiles)
summarizeProfiles(example_profiles)
}
