% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruction.R
\name{reconstructDominant}
\alias{reconstructDominant}
\title{Reconstruct Reference Isoform (Deprecated)}
\usage{
reconstructDominant(comparator_exons, events)
}
\arguments{
\item{comparator_exons}{Data frame of comparator exons with columns:
chr, exon_start, exon_end, strand, gene_id, transcript_id.}

\item{events}{Tibble of detected events from \code{\link[=detectEvents]{detectEvents()}}.}
}
\value{
Data frame of reconstructed reference exons.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{reconstructDominant()} was renamed to \code{\link[=reconstructReference]{reconstructReference()}} to better
reflect the reference/comparator terminology used throughout the package.
}
\keyword{internal}
